requirements.txt
fastapi>=0.104.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
mcp
openai
requests
python-dotenv
typing-extensions